+++ designer/client/src/components/sidePanels/SidePanel.tsx
@@ -7,7 +7,7 @@ import { isLeftPanelOpened, isRightPanelOpened } from "../../reducers/selectors/
 import { togglePanel } from "../../actions/nk";
 import { useGraph } from "../graph/GraphContext";
 import { Graph } from "../graph/Graph";
-import styled from "@emotion/styled";
+import { styled } from "@mui/material";
 
 const SCROLL_THUMB_SIZE = 8;
 const SIDEBAR_WIDTH = 290;
@@ -56,7 +56,7 @@ function checkRightSide(props: ScrollToggle) {
     }
 }
 
-const ScrollToggleChild = styled.div((props: ScrollToggle) => ({
+const ScrollToggleChild = styled("div")((props: ScrollToggle) => ({
     width: PANEL_WIDTH,
     boxSizing: "border-box",
     minHeight: "100%",
@@ -66,7 +66,7 @@ const ScrollToggleChild = styled.div((props: ScrollToggle) => ({
     alignItems: props.side === PanelSide.Left ? "flex-start" : "flex-end",
 }));
 
-const ScrollToggle = styled.div((props: ScrollToggle) => ({
+const ScrollToggle = styled("div")((props: ScrollToggle) => ({
     pointerEvents: "none",
     userSelect: "none",
     width: PANEL_WIDTH,